5.3.x

+++ spring-web/src/main/java/org/springframework/web/multipart/support/StandardServletMultipartResolver.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2002-2017 the original author or authors.
+ * Copyright 2002-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -21,6 +21,8 @@ import javax.servlet.http.Part;
 
 import org.apache.commons.logging.LogFactory;
 
+import org.springframework.http.MediaType;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartException;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.multipart.MultipartResolver;
@@ -31,6 +33,15 @@ import org.springframework.web.multipart.MultipartResolver;
  * To be added as "multipartResolver" bean to a Spring DispatcherServlet context,
  * without any extra configuration at the bean level (see below).
  *
+ * <p>This resolver variant uses your Servlet container's multipart parser as-is,
+ * potentially exposing the application to container implementation differences.
+ * See {@link org.springframework.web.multipart.commons.CommonsMultipartResolver}
+ * for an alternative implementation using a local Commons FileUpload library
+ * within the application, providing maximum portability across Servlet containers.
+ * Also, see this resolver's configuration option for
+ * {@link #setStrictServletCompliance strict Servlet compliance}, narrowing the
+ * applicability of Spring's {@link MultipartHttpServletRequest} to form data only.
+ *
  * <p><b>Note:</b> In order to use Servlet 3.0 based multipart parsing,
  * you need to mark the affected servlet with a "multipart-config" section in
  * {@code web.xml}, or with a {@link javax.servlet.MultipartConfigElement}
@@ -40,15 +51,29 @@ import org.springframework.web.multipart.MultipartResolver;
  * storage locations need to be applied at that servlet registration level;
  * Servlet 3.0 does not allow for them to be set at the MultipartResolver level.
  *
+ * <pre class="code">
+ * public class AppInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
+ *	 // ...
+ *	 &#064;Override
+ *	 protected void customizeRegistration(ServletRegistration.Dynamic registration) {
+ *     // Optionally also set maxFileSize, maxRequestSize, fileSizeThreshold
+ *     registration.setMultipartConfig(new MultipartConfigElement("/tmp"));
+ *   }
+ * }
+ * </pre>
+ *
  * @author Juergen Hoeller
  * @since 3.1
  * @see #setResolveLazily
+ * @see #setStrictServletCompliance
  * @see HttpServletRequest#getParts()
  * @see org.springframework.web.multipart.commons.CommonsMultipartResolver
  */
 public class StandardServletMultipartResolver implements MultipartResolver {
 
 	private boolean resolveLazily = false;
+
+	private boolean strictServletCompliance = false;
 
 
 	/**
@@ -58,20 +83,38 @@ public class StandardServletMultipartResolver implements MultipartResolver {
 	 * corresponding exceptions at the time of the {@link #resolveMultipart} call.
 	 * Switch this to "true" for lazy multipart parsing, throwing parse exceptions
 	 * once the application attempts to obtain multipart files or parameters.
+	 * @since 3.2.9
 	 */
 	public void setResolveLazily(boolean resolveLazily) {
 		this.resolveLazily = resolveLazily;
 	}
 
+	/**
+	 * Specify whether this resolver should strictly comply with the Servlet
+	 * specification, only kicking in for "multipart/form-data" requests.
+	 * <p>Default is "false", trying to process any request with a "multipart/"
+	 * content type as far as the underlying Servlet container supports it
+	 * (which works on e.g. Tomcat but not on Jetty). For consistent portability
+	 * and in particular for consistent custom handling of non-form multipart
+	 * request types outside of Spring's {@link MultipartResolver} mechanism,
+	 * switch this flag to "true": Only "multipart/form-data" requests will be
+	 * wrapped with a {@link MultipartHttpServletRequest} then; other kinds of
+	 * requests will be left as-is, allowing for custom processing in user code.
+	 * <p>Note that Commons FileUpload and therefore
+	 * {@link org.springframework.web.multipart.commons.CommonsMultipartResolver}
+	 * supports any "multipart/" request type. However, it restricts processing
+	 * to POST requests which standard Servlet multipart parsers might not do.
+	 * @since 5.3.9
+	 */
+	public void setStrictServletCompliance(boolean strictServletCompliance) {
+		this.strictServletCompliance = strictServletCompliance;
+	}
+
 
 	@Override
 	public boolean isMultipart(HttpServletRequest request) {
-		// Same check as in Commons FileUpload...
-		if (!"post".equals(request.getMethod().toLowerCase())) {
-			return false;
-		}
-		String contentType = request.getContentType();
-		return (contentType != null && contentType.toLowerCase().startsWith("multipart/"));
+		return StringUtils.startsWithIgnoreCase(request.getContentType(),
+				(this.strictServletCompliance ? MediaType.MULTIPART_FORM_DATA_VALUE : "multipart/"));
 	}
 
 	@Override
@@ -81,17 +124,20 @@ public class StandardServletMultipartResolver implements MultipartResolver {
 
 	@Override
 	public void cleanupMultipart(MultipartHttpServletRequest request) {
-		// To be on the safe side: explicitly delete the parts,
-		// but only actual file parts (for Resin compatibility)
-		try {
-			for (Part part : request.getParts()) {
-				if (request.getFile(part.getName()) != null) {
-					part.delete();
+		if (!(request instanceof AbstractMultipartHttpServletRequest) ||
+				((AbstractMultipartHttpServletRequest) request).isResolved()) {
+			// To be on the safe side: explicitly delete the parts,
+			// but only actual file parts (for Resin compatibility)
+			try {
+				for (Part part : request.getParts()) {
+					if (request.getFile(part.getName()) != null) {
+						part.delete();
+					}
 				}
 			}
-		}
-		catch (Throwable ex) {
-			LogFactory.getLog(getClass()).warn("Failed to perform cleanup of multipart items", ex);
+			catch (Throwable ex) {
+				LogFactory.getLog(getClass()).warn("Failed to perform cleanup of multipart items", ex);
+			}
 		}
 	}
